move tests to correct suite, write test for collection exercise id returing no cases
